--- src/main/java/bumblebee/xchangepass/domain/ExchangeRate/repository/ExchangeRateTempRepository.java
@@ -1,8 +1,0 @@
-package bumblebee.xchangepass.domain.ExchangeRate.repository;
-
-import bumblebee.xchangepass.domain.ExchangeRate.entity.ExchangeRateTemp;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ExchangeRateTempRepository extends JpaRepository<ExchangeRateTemp, Long> {
-
-}
