--- src/main/java/bumblebee/xchangepass/domain/ExchangeTransaction/entitiy/TransactionStatus.java
@@ -1,5 +1,0 @@
-package bumblebee.xchangepass.domain.ExchangeTransaction.entitiy;
-
-public enum TransactionStatus {
-    PENDING, COMPLETED, FAILED
-}
