--- src/main/java/bumblebee/xchangepass/domain/ExchangeTransaction/repository/ExchangeTransactionRepository.java
@@ -1,8 +1,0 @@
-package bumblebee.xchangepass.domain.ExchangeTransaction.repository;
-
-import bumblebee.xchangepass.domain.ExchangeTransaction.entitiy.ExchangeTransaction;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ExchangeTransactionRepository extends JpaRepository<ExchangeTransaction, Long> {
-
-}
