+++ src/main/java/bumblebee/xchangepass/domain/exchangeRate/repository/exchangeTemp/ExchangeRepositoryImpl.java
@@ -1,0 +1,54 @@
+package bumblebee.xchangepass.domain.exchangeRate.repository.exchangeTemp;
+
+import bumblebee.xchangepass.global.error.ErrorCode;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import jakarta.persistence.PersistenceException;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class ExchangeRepositoryImpl implements ExchangeRepositoryCustom {
+
+    @PersistenceContext
+    private EntityManager entityManager;
+
+    public void renameTable(String oldTableName, String newTableName) {
+        try {
+            entityManager.createNativeQuery("ALTER TABLE " + oldTableName + " RENAME TO " + newTableName).executeUpdate();
+        }catch (PersistenceException e) {
+            throw ErrorCode.EXCHANGE_TABLE_RENAME_FAIL.commonException();
+        }
+    }
+
+    public void createTempTable() {
+        try {
+            String sql = "CREATE TABLE IF NOT EXISTS exchange_rate_temp (" +
+                    "id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY, " +
+                    "base_currency VARCHAR(255) NOT NULL, " +
+                    "exchange_rates JSONB, " +
+                    "updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP)";
+            entityManager.createNativeQuery(sql).executeUpdate();
+            // ✅ GIN 인덱스 추가
+            String indexSql = "CREATE INDEX IF NOT EXISTS exchange_rate_temp_jsonb_idx " +
+                    "ON exchange_rate_temp USING GIN (exchange_rates)";
+            entityManager.createNativeQuery(indexSql).executeUpdate();
+        }catch (PersistenceException e) {
+            throw ErrorCode.EXCHANGE_TABLE_RENAME_FAIL.commonException();
+        }
+    }
+
+    public boolean isTableExist(String tableName) {
+        return ((Number) entityManager.createNativeQuery(
+                        "SELECT COUNT(*) FROM information_schema.tables WHERE table_name = ?")
+                .setParameter(1, tableName)
+                .getSingleResult()).intValue() > 0;
+    }
+
+    public void dropTableIfExists(String tableName) {
+        try {
+            entityManager.createNativeQuery("DROP TABLE IF EXISTS " + tableName).executeUpdate();
+        } catch (PersistenceException e) {
+            throw ErrorCode.EXCHANGE_TABLE_RENAME_FAIL.commonException();
+        }
+    }
+}
