+++ src/main/java/bumblebee/xchangepass/domain/exchangeTransaction/entitiy/TransactionStatus.java
@@ -1,0 +1,5 @@
+package bumblebee.xchangepass.domain.exchangeTransaction.entitiy;
+
+public enum TransactionStatus {
+    PENDING, COMPLETED, FAILED
+}
