+++ src/main/java/bumblebee/xchangepass/domain/wallet/service/WalletService.java
@@ -1,6 +1,6 @@
 package bumblebee.xchangepass.domain.wallet.service;
 
-import bumblebee.xchangepass.domain.ExchangeRate.service.ExchangeService;
+import bumblebee.xchangepass.domain.exchangeRate.service.ExchangeService;
 import bumblebee.xchangepass.domain.user.entity.User;
 import bumblebee.xchangepass.domain.user.repository.UserRepository;
 import bumblebee.xchangepass.domain.wallet.dto.request.WalletInOutRequest;
